requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
apscheduler==3.10.4
asyncio==3.4.3